goit-pyweb-hw-09/pyproject.toml
[tool.poetry]
name = "goit-pyweb-hw-09"
version = "0.1.0"
description = ""
authors = ["MaxRevan <gladkovnissan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
scrapy = "^2.11.2"
mongoengine = "^0.29.1"
python-dotenv = "^1.0.1"
pymongo = "^4.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
